# Conflicts:
#	src/main/java/uk/gov/justice/hmpps/prison/api/model/PrisonerBookingSummary.java
#	src/main/java/uk/gov/justice/hmpps/prison/api/resource/BookingResource.java
#	src/main/java/uk/gov/justice/hmpps/prison/repository/jpa/model/AgencyLocation.java
#	src/main/java/uk/gov/justice/hmpps/prison/service/BookingService.java
#	src/main/java/uk/gov/justice/hmpps/prison/service/transformers/OffenderBookingTransformer.java
#	src/test/java/uk/gov/justice/hmpps/prison/api/resource/impl/BookingResourceIntTest.java
#	src/test/java/uk/gov/justice/hmpps/prison/service/BookingServiceTest.java
#	src/test/resources/uk/gov/justice/hmpps/prison/api/resource/impl/bookings_by_nomsId.json
#	src/test/resources/uk/gov/justice/hmpps/prison/api/resource/impl/lei_bookings.json

+++ src/main/java/uk/gov/justice/hmpps/prison/api/model/PrivilegeSummary.java
@@ -48,7 +48,7 @@ public class PrivilegeSummary {
 
     @ApiModelProperty(required = true, value = "The number of days since last review.", example = "35", position = 5)
     @NotNull
-    private Integer daysSinceReview;
+    private Long daysSinceReview;
 
     @ApiModelProperty(value = "All IEP detail entries for the offender (most recent first).", position = 6)
     private List<PrivilegeDetail> iepDetails;
